Don't allow deleting category when there is item in it

+++ src/main/java/com/wusly/backendmenu/service/CategoryService.java
@@ -4,7 +4,10 @@ import com.wusly.backendmenu.domain.category.AddCategoryCommand;
 import com.wusly.backendmenu.domain.category.Category;
 import com.wusly.backendmenu.domain.category.CategoryResponse;
 import com.wusly.backendmenu.error.NotFoundException;
+import com.wusly.backendmenu.error.NotPermittedException;
 import com.wusly.backendmenu.repository.CategoryRepository;
+import com.wusly.backendmenu.service.item.ItemService;
+import com.wusly.backendmenu.service.item.ItemServiceHelper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,6 +20,7 @@ import java.util.UUID;
 public class CategoryService {
     private final CategoryRepository categoryRepository;
     private final RestaurantService restaurantService;
+    private final ItemServiceHelper itemServiceHelper;
 
     public boolean existsById(UUID id) {
         return categoryRepository.existsById(id);
@@ -55,9 +59,11 @@ public class CategoryService {
     public void delete(UUID id, String email) {
         var restaurant = restaurantService.getRestaurantByEmail(email);
         var category = categoryRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("Catehory with Id: %s not found!"));
+                .orElseThrow(() -> new NotFoundException("Category with Id: %s not found!"));
+        if(itemServiceHelper.existsByCategoryId(id))
+            throw new NotPermittedException("Before deleting the category you should delete the items In category!");
         if(!restaurant.getId().equals(category.getRestaurantId()))
-            throw new NotFoundException("Catehory with Id: %s not found!");
+            throw new NotFoundException("Category with Id: %s not found!");
         categoryRepository.delete(category);
     }
 }
